split layer-examples by category
