+++ examples/main/src/layer-examples/index.js
@@ -1,0 +1,9 @@
+import coreLayerExamples from './core-layers';
+import deprecatedLayerExamples from './deprecated-layers';
+import sampleLayerExamples from './sample-layers';
+
+export default {
+  ...coreLayerExamples,
+  ...deprecatedLayerExamples,
+  ...sampleLayerExamples
+};
